refactor: divide Update template component
