+++ src/components/record/RecordInfo.tsx
@@ -1,0 +1,11 @@
+import UpdateTemplateTemplateContent from '@/components/Template/UpdateTemplateContent';
+import RecordDetailTitle from '@/components/UpdateTemplate/UpdateTemplateTitle';
+
+export default function RecordDetailTemplate() {
+  return (
+    <>
+      <RecordDetailTitle />
+      <UpdateTemplateTemplateContent />
+    </>
+  );
+}
